fix(constant-stream): move outside of node_modules
npm 5 correctly not package all node_modules entries, therefore breaking
xo-server package.
Stop doing this for the moment.
