+++ src/constant-stream.js
